Fix all compile time warnings.
Just being sloppy there, best practice is to not have *any* warnings, so this cleans up all of those.

Found an actual bug doing this- the Nektra return parameter is poorly defined as long and long long on x64.  Should be a pointer value for function named put_PointerVal, but their APIs pretty much suck.  On x64 games this could have truncated anything built above the 4G boundary.

Fixes some BSTR conversions as well, since they did not need to be ComPtr.

+++ DeviarePlugin/InProc_DX9.cpp
@@ -160,8 +160,8 @@ void DrawStereoOnGame(IDirect3DDevice9* device, IDirect3DSurface9* surface, IDir
 {
 	D3DSURFACE_DESC bufferDesc;
 	back->GetDesc(&bufferDesc);
-	RECT backBufferRect = { 0, 0, bufferDesc.Width, bufferDesc.Height };
-	RECT stereoImageRect = { 0, 0, bufferDesc.Width * 2, bufferDesc.Height };
+	RECT backBufferRect = { 0, 0, (LONG) bufferDesc.Width, (LONG) bufferDesc.Height };
+	RECT stereoImageRect = { 0, 0, (LONG) bufferDesc.Width * 2, (LONG) bufferDesc.Height };
 
 	int insetW = 300;
 	int insetH = (int)(300.0 * stereoImageRect.bottom / stereoImageRect.right);
